+++ packages/reactabular-resolve/CHANGELOG.md
@@ -1,3 +1,8 @@
+2.0.2 / 2016-08-17
+==================
+
+  * Bug fix - Make sure `resolve` does not crash if `rows` aren't provided. It will return an empty array in that case.
+
 2.0.0 / 2016-08-16
 ==================